pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canne"
version = "0.1.0"
description = "Système d'Assistance Intelligent pour Malvoyants"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "Canne Project Contributors", email = "info@example.com"}
]
dependencies = [
    "RPi.GPIO>=0.7.0",
    "numpy>=1.19.0",
    "opencv-python>=4.5.0",
    "tensorflow>=2.5.0",
    "gtts>=2.2.3",
    "playsound>=1.3.0",
    "pygame>=2.0.1",
    "pillow>=8.0.0",
    "pyyaml>=5.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "pytest-cov>=2.12.1",
    "flake8>=3.9.2",
    "black>=21.5b2"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=src"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
''' 